refactor(colab): extract ColorizedMask/ImagePreview, use presigned image URLs
Moves the label-recoloring canvas component and the raw/annotated preview
pane out of ImageViewer.tsx into a shared src/components/colab/ImagePreview.tsx,
so the upcoming dataset overview (colab-rework-plan.md F4) can reuse them
without depending on the legacy session dashboard.

Also swaps the public artifact-manager URL construction for presigned
get_file calls (withStageRetry from datasetApi.ts) when loading the selected
image and its latest mask, per plan section 6 F4.

Verification note: `pnpm exec tsc --noEmit` cannot run in this repo in its
current state (typescript@3.9.10 is pinned everywhere, including inside
react-scripts' own node_modules, and cannot parse modern .d.ts syntax shipped
by @headlessui/react@2.2.0 and @floating-ui/utils@0.2.9 — a parser failure,
not a tsconfig issue; skipLibCheck does not help). This is pre-existing and
unrelated to this change. Used @babel/core + @babel/preset-typescript
programmatically as a syntax-check substitute; both changed files parse
cleanly.
