+++ src/components/colab/ImageViewer.tsx
@@ -1,95 +1,6 @@
 import React, { useState, useEffect, useCallback } from 'react';
-
-const ColorizedMask = ({
-  src,
-  alt,
-  className,
-  onError,
-}: {
-  src: string;
-  alt: string;
-  className: string;
-  onError: (e: any) => void;
-}) => {
-  const [dataUrl, setDataUrl] = useState<string | null>(null);
-
-  useEffect(() => {
-    let active = true;
-
-    const loadAndColorize = () => {
-      const img = new Image();
-      img.crossOrigin = 'anonymous';
-      img.onload = () => {
-        if (!active) return;
-        const canvas = document.createElement('canvas');
-        canvas.width = img.width;
-        canvas.height = img.height;
-        const ctx = canvas.getContext('2d');
-        if (!ctx) return;
-
-        ctx.drawImage(img, 0, 0);
-
-        let imgData;
-        try {
-          imgData = ctx.getImageData(0, 0, canvas.width, canvas.height);
-        } catch (e) {
-          console.warn("CORS error getting image data, using raw image", e);
-          setDataUrl(src);
-          return;
-        }
-
-        const data = imgData.data;
-
-        const palette = [
-          [37, 99, 235],
-          [147, 51, 234],
-          [219, 39, 119],
-          [22, 163, 74],
-          [234, 88, 12],
-          [220, 38, 38],
-          [202, 138, 4],
-          [6, 182, 212],
-          [139, 92, 246],
-          [244, 63, 94],
-        ];
-
-        for (let i = 0; i < data.length; i += 4) {
-          const r = data[i];
-          const g = data[i + 1];
-          const labelId = (r << 8) | g;
-
-          if (labelId === 0) {
-            data[i] = 0; data[i + 1] = 0; data[i + 2] = 0; data[i + 3] = 255;
-          } else {
-            const color = palette[(labelId - 1) % palette.length];
-            data[i] = color[0]; data[i + 1] = color[1]; data[i + 2] = color[2]; data[i + 3] = 255;
-          }
-        }
-        ctx.putImageData(imgData, 0, 0);
-        setDataUrl(canvas.toDataURL('image/png'));
-      };
-      img.onerror = onError;
-      img.src = src;
-    };
-
-    loadAndColorize();
-
-    return () => { active = false; };
-  }, [src, onError]);
-
-  if (!dataUrl) {
-    return (
-      <div className={`flex items-center justify-center bg-black/5 rounded-lg ${className}`} style={{ minHeight: '300px' }}>
-        <svg className="w-8 h-8 animate-spin text-purple-600" fill="none" viewBox="0 0 24 24">
-          <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-          <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-        </svg>
-      </div>
-    );
-  }
-
-  return <img src={dataUrl} alt={alt} className={className} />;
-};
+import { ImagePreview } from './ImagePreview';
+import { withStageRetry } from './datasetApi';
 
 export interface SplitInfo {
   applied: boolean;
@@ -349,15 +260,26 @@ const ImageViewer: React.FC<ImageViewerProps> = ({
 
     const loadImageUrls = async () => {
       try {
-        const artifactAlias = dataArtifactId.split('/').pop();
         const baseName = selectedImage.substring(0, selectedImage.lastIndexOf('.')) || selectedImage;
         const annotated = isAnnotated(selectedImage);
         const imageInArtifact = cloudImages.has(selectedImage);
         const imageFolder = getImageFolder(selectedImage);
 
         if (imageInArtifact) {
-          const rawUrl = `${serverUrl}/bioimage-io/artifacts/${artifactAlias}/files/${imageFolder}/${baseName}.png`;
-          setImageUrl(rawUrl);
+          try {
+            const rawUrl = await withStageRetry(() =>
+              artifactManager.get_file({
+                artifact_id: dataArtifactId,
+                file_path: `${imageFolder}/${baseName}.png`,
+                stage: true,
+                _rkwargs: true,
+              }),
+            );
+            setImageUrl(rawUrl);
+          } catch (error) {
+            console.error('Failed to get presigned image URL:', error);
+            setImageUrl('');
+          }
         } else {
           let localImageLoaded = false;
 
@@ -405,8 +327,20 @@ const ImageViewer: React.FC<ImageViewerProps> = ({
         if (annotated) {
           const maskFolder = label ? `masks_${label}` : 'annotations';
           const relPath = findAnnotationRelPath(selectedImage) || `${baseName}.png`;
-          const annUrl = `${serverUrl}/bioimage-io/artifacts/${artifactAlias}/files/${maskFolder}/${relPath}`;
-          setAnnotationUrl(annUrl);
+          try {
+            const annUrl = await withStageRetry(() =>
+              artifactManager.get_file({
+                artifact_id: dataArtifactId,
+                file_path: `${maskFolder}/${relPath}`,
+                stage: true,
+                _rkwargs: true,
+              }),
+            );
+            setAnnotationUrl(annUrl);
+          } catch (error) {
+            console.error('Failed to get presigned annotation URL:', error);
+            setAnnotationUrl('');
+          }
         } else {
           setAnnotationUrl('');
         }
@@ -416,7 +350,7 @@ const ImageViewer: React.FC<ImageViewerProps> = ({
     };
 
     loadImageUrls();
-  }, [selectedImage, dataArtifactId, label, serverUrl, annotationsList, dataSourceType, imageFolderHandle, executeCode, annotationURL, server, cloudImages, getImageFolder]);
+  }, [selectedImage, dataArtifactId, label, artifactManager, annotationsList, dataSourceType, imageFolderHandle, executeCode, annotationURL, server, cloudImages, getImageFolder]);
 
   const totalImages = imageList.length;
   const annotatedImages = imageList.filter(img => isAnnotated(img)).length;
@@ -689,42 +623,13 @@ const ImageViewer: React.FC<ImageViewerProps> = ({
     return (
       <>
         <div className="flex-1 flex items-center justify-center p-6 bg-gray-50 overflow-auto">
-          {viewMode === 'annotated' && !hasAnnotation ? (
-            <div className="text-center text-gray-400">
-              <svg className="w-16 h-16 mx-auto mb-3" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1} d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
-              </svg>
-              <p className="text-sm">No annotation for this image yet.</p>
-            </div>
-          ) : viewMode === 'raw' && imageUrl ? (
-            <img
-              src={imageUrl}
-              alt={selectedImage}
-              className="max-w-full max-h-full object-contain rounded-lg shadow-lg"
-              onError={(e) => {
-                (e.target as HTMLImageElement).src = 'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDAwIiBoZWlnaHQ9IjMwMCIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cmVjdCB3aWR0aD0iNDAwIiBoZWlnaHQ9IjMwMCIgZmlsbD0iI2YzZjRmNiIvPjx0ZXh0IHg9IjUwJSIgeT0iNTAlIiBmb250LWZhbWlseT0iQXJpYWwiIGZvbnQtc2l6ZT0iMTgiIGZpbGw9IiM5Y2EzYWYiIHRleHQtYW5jaG9yPSJtaWRkbGUiIGR5PSIuM2VtIj5JbWFnZSBub3QgYXZhaWxhYmxlPC90ZXh0Pjwvc3ZnPg==';
-              }}
-            />
-          ) : viewMode === 'annotated' && annotationUrl ? (
-            <ColorizedMask
-              src={annotationUrl}
-              alt={`${selectedImage} (annotated)`}
-              className="max-w-full max-h-full object-contain rounded-lg shadow-lg"
-              onError={(e) => {
-                const target = e.target || e;
-                if (target && typeof target === 'object' && 'src' in target) {
-                  (target as HTMLImageElement).src = 'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDAwIiBoZWlnaHQ9IjMwMCIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cmVjdCB3aWR0aD0iNDAwIiBoZWlnaHQ9IjMwMCIgZmlsbD0iI2YzZjRmNiIvPjx0ZXh0IHg9IjUwJSIgeT0iNTAlIiBmb250LWZhbWlseT0iQXJpYWwiIGZvbnQtc2l6ZT0iMTgiIGZpbGw9IiM5Y2EzYWYiIHRleHQtYW5jaG9yPSJtaWRkbGUiIGR5PSIuM2VtIj5Bbm5vdGF0aW9uIG5vdCBhdmFpbGFibGU8L3RleHQ+PC9zdmc+';
-                }
-              }}
-            />
-          ) : (
-            <div className="text-center text-gray-400">
-              <svg className="w-16 h-16 mx-auto mb-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1} d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z" />
-              </svg>
-              <p>Loading image...</p>
-            </div>
-          )}
+          <ImagePreview
+            viewMode={viewMode}
+            imageUrl={imageUrl}
+            annotationUrl={annotationUrl}
+            hasAnnotation={hasAnnotation}
+            alt={selectedImage}
+          />
         </div>
 
         <div className="bg-white border-t border-gray-200 px-6 py-3 flex-shrink-0">
